refactor: move the useOrderStore to store folder and use the correct path for it in the components
