+++ src/components/admin/OrderForm.jsx
@@ -1,5 +1,5 @@
 
-import useOrderStore from "../../store/admin/useOrderStore";
+import useOrderStore from "../../store/useOrderStore";
 import Button from "../common/order/Button";
 import { useNavigate, Link } from 'react-router-dom';
 import toast from 'react-hot-toast';
