+++ src/components/admin/order-from-sections/Location.jsx
@@ -1,5 +1,4 @@
-
-import useOrderStore from "../../../store/admin/useOrderStore";
+import useOrderStore from "../../../store/useOrderStore";
 import Map from "../../common/order/Map";
 import { LuMapPin, LuNavigation } from "react-icons/lu";
