+++ src/components/admin/order-from-sections/ServiceInfo.jsx
@@ -2,7 +2,7 @@
 import { MdOutlineSettingsSuggest } from "react-icons/md";
 import Dropdown from "../../common/Dropdown"; 
 import { useEffect, useState } from "react";
-import useOrderStore from "../../../store/admin/useOrderStore";
+import useOrderStore from "../../../store/useOrderStore";
 
 export default function ServiceInfo() {
