+++ src/components/common/order/CancelOrder.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 // import { X, AlertTriangle } from "lucide-react";
-import useOrderStore from "../../../store/admin/useOrderStore";
+import useOrderStore from "../../../store/useOrderStore";
 import toast from "react-hot-toast";
 import { useTranslation } from "react-i18next";
 import { FiX } from "react-icons/fi";
